13/requirements.txt
line-bot-sdk==3.5.0
requests==2.31.0
flask==3.0.0